// resource/generated/models/OfferItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BarbarianFoodAPI
 * example
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OfferItem
 */
export interface OfferItem {
    /**
     * 
     * @type {string}
     * @memberof OfferItem
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OfferItem
     */
    title?: string;
    /**
     * 
     * @type {number}
     * @memberof OfferItem
     */
    kcalRangeBottom?: number;
    /**
     * 
     * @type {number}
     * @memberof OfferItem
     */
    kcalRangeTop?: number;
    /**
     * 
     * @type {number}
     * @memberof OfferItem
     */
    cost?: number;
}

export function OfferItemFromJSON(json: any): OfferItem {
    return OfferItemFromJSONTyped(json, false);
}

export function OfferItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): OfferItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'title': !exists(json, 'title') ? undefined : json['title'],
        'kcalRangeBottom': !exists(json, 'kcalRangeBottom') ? undefined : json['kcalRangeBottom'],
        'kcalRangeTop': !exists(json, 'kcalRangeTop') ? undefined : json['kcalRangeTop'],
        'cost': !exists(json, 'cost') ? undefined : json['cost'],
    };
}

export function OfferItemToJSON(value?: OfferItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'kcalRangeBottom': value.kcalRangeBottom,
        'kcalRangeTop': value.kcalRangeTop,
        'cost': value.cost,
    };
}