// resource/generated/models/UserItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BarbarianFoodAPI
 * example
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserItem
 */
export interface UserItem {
    /**
     * 
     * @type {string}
     * @memberof UserItem
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UserItem
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UserItem
     */
    surname?: string;
    /**
     * 
     * @type {string}
     * @memberof UserItem
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserItem
     */
    activeSubscription?: string;
}

export function UserItemFromJSON(json: any): UserItem {
    return UserItemFromJSONTyped(json, false);
}

export function UserItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'surname': !exists(json, 'surname') ? undefined : json['surname'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'activeSubscription': !exists(json, 'activeSubscription') ? undefined : json['activeSubscription'],
    };
}

export function UserItemToJSON(value?: UserItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'surname': value.surname,
        'email': value.email,
        'activeSubscription': value.activeSubscription,
    };
}